+++ ff-gtk/Main.hs
@@ -19,7 +19,8 @@ import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import qualified Data.Text as Text
 import qualified Data.Vector as Vector
-import FF (fromRgaM, getDataDir, loadTasks, noDataDirectoryMessage)
+import FF (fromRgaM, getDataDir, filterTasksByStatus, loadAllNotes,
+          noDataDirectoryMessage, toNoteView)
 import FF.Config (loadConfig)
 import FF.Types
   ( Entity (Entity),
@@ -123,7 +124,10 @@ main = do
 
 initiallyLoadActiveTasks :: StorageFS.Handle -> Producer Event IO ()
 initiallyLoadActiveTasks storage = do
-  activeTasks <- lift $ runStorage storage $ loadTasks Active
+  activeTasks <- lift $ runStorage storage $ do
+    notes <- loadAllNotes
+    let filtered = filterTasksByStatus Active notes
+    traverse toNoteView filtered
   each $ map UpsertTask activeTasks
 
 getDataDirOrFail :: IO FilePath